requirements.txt
sqlalchemy
pydantic
starlette
fastapi
uvicorn
passlib
bcrypt
google-auth-oauthlib
google-api-python-client
python-dotenv